fix(auth0): use datatools-server to retrieve user profile from id token
This (in conjunction with changes in datatools-server) allows new Auth0 tenants to work with Data
Tools. Instead of fetching the user profile from Auth0, we fetch it directly from the
datatools-server API. The server decodes the JWT token into the user profile and re-maps certain
fields (e.g., the scoped app_metadata and the user_id fields) to keys the front end expects.

fixes #207